Add files via upload
PLAGIARISM DETECTOR

What is the Project About
This project is a Plagiarism Detector Website. You paste some text into it, click a button, and it tells you if the text is copied or original. It uses a smart program (machine learning) to check the text.



What Problem Does It Fix
Many people copy content from the internet or other sources. It’s hard to find out if something is copied just by reading. This project helps by checking the text automatically and quickly tells if it’s plagiarized or not.



Why Did We Make This Project
We made this project to:

* Help teachers and students check for copied work.
* Learn how to use web tools and machine learning together.
* Create a useful tool that saves time and effort.